execpeersite/requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
certifi==2018.8.24
chardet==3.0.4
colorama==0.3.9
decorator==4.3.0
dj-database-url==0.5.0
Django==2.1.1
django-heroku==0.3.1
gunicorn==19.9.0
idna==2.7
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
more-itertools==4.3.0
oauthlib==2.1.0
parso==0.3.1
pickleshare==0.7.4
pluggy==0.7.1
prompt-toolkit==1.0.15
psycopg2==2.7.5
py==1.6.0
Pygments==2.2.0
pytest==3.8.1
python-decouple==3.1
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
whitenoise==4.1
